requirements.txt
Flask
ldap3
requests
python-dotenv